Fix: Format specifier warnings for uint32_t
Fixed all -Wformat warnings by changing %u to %lu with cast to
(unsigned long) for uint32_t variables.

Changes:
- frame_queue.c (lines 293, 350): buffer_size formatting
- camera_threads.c (lines 177, 236, 247, 336, 365, 379):
  error_count, frame_count, packet_count, total_bytes formatting
- camera_app_main.c (line 411): frame_count formatting

Build now completes with zero warnings in security_camera files.

+++ apps/examples/security_camera/camera_app_main.c
@@ -408,7 +408,8 @@ int main(int argc, FAR char *argv[])
     }
   else
     {
-      LOG_INFO("Main loop ended, total frames: %u", frame_count);
+      LOG_INFO("Main loop ended, total frames: %lu",
+               (unsigned long)frame_count);
     }
 
   /* Cleanup */
